resume_parser/requirements.txt
spacy==3.7.2
python-docx==1.1.0
pdfminer.six==20221105
pytest==8.2.1
scikit-learn==1.5.0
pandas==2.2.2
